Fix signed-unsigned comparisons. (#99)
When compiling under gcc7, I see warnings about comparison of signed
and unsigned integer types. This change enables the same warnings for
clang, and fixes the current sites.

+++ bftengine/src/communication/TlsTCPCommunication.cpp
@@ -509,7 +509,7 @@ class AsyncTlsConnection : public
     // the certificate must have node id, as we put it in OU field on creation.
     // since we use pinning we must know who is the remote peer.
     // peerIdPrefixLength stands for the length of 'OU=' substring
-    size_t peerIdPrefixLength = 3;
+    int peerIdPrefixLength = 3;
     std::regex r("OU=\\d*", std::regex_constants::icase);
     std::smatch sm;
     regex_search(subject, sm, r);
